requirements.txt
pyrogram
tgcrypto
loguru
lightdb==1.1.1
gitpython